Moved GamesTable Out of Beta.
