+++ agagd/agagd_core/views/frontpage.py
@@ -5,7 +5,8 @@ import datetime
 import agagd_core.models as agagd_models
 
 # Django Tables Imports
-from agagd_core.tables.beta import GamesTable, PlayersTournamentTable
+from agagd_core.tables.beta import PlayersTournamentTable
+from agagd_core.tables.games import GamesTable
 from agagd_core.tables.top_players import TopDanTable, TopKyuTable
 from agagd_core.tables.tournaments import TournamentsTable
 from django.shortcuts import render
